requirements.txt
django
djangorestframework
django-filter
sphinx
rinohtype
path